Adicionando DTO (Data transfer object)

+++ src/main/java/com/estudo/estudofaculdade/services/ProdutoService.java
@@ -2,12 +2,16 @@ package com.estudo.estudofaculdade.services;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.estudo.estudofaculdade.model.Produto;
+import com.estudo.estudofaculdade.model.exception.ResourceNotFoundException;
 import com.estudo.estudofaculdade.repository.ProdutoRepository;
+import com.estudo.estudofaculdade.shared.ProdutoDTO;
 
 @Service
 public class ProdutoService {
@@ -20,8 +24,13 @@ public class ProdutoService {
      * 
      * @return lista de produtos.
      */
-    public List<Produto> obterTodos() {
-        return produtoRepository.findAll();
+    public List<ProdutoDTO> obterTodos() {
+
+        List<Produto> produtos = produtoRepository.findAll();
+
+        return produtos.stream()
+                .map(produto -> new ModelMapper().map(produto, ProdutoDTO.class))
+                .collect(Collectors.toList());
     }
 
     /**
@@ -30,8 +39,15 @@ public class ProdutoService {
      * @param id do produto que será localizado
      * @return Retona um produto caso seja encontrado.
      */
-    public Optional<Produto> obterPorId(Integer id) {
-        return produtoRepository.findById(id);
+    public Optional<ProdutoDTO> obterPorId(Integer id) {
+        Optional<Produto> produto = produtoRepository.findById(id);
+
+        if (produto.isEmpty()) {
+            throw new ResourceNotFoundException("Produto com id: " + id + "não encontrado");
+        }
+
+        ProdutoDTO dto = new ModelMapper().map(produto.get(), ProdutoDTO.class);
+        return Optional.of(dto);
     }
 
     /**
@@ -40,9 +56,24 @@ public class ProdutoService {
      * @param produto que será adicionado.
      * @return Retorna o produto que foi adicionado na lista.
      */
-    public Produto adicionar(Produto produto) {
-        // Poderia ter alguma regra de negocio aqui para validar o produto.
-        return produtoRepository.save(produto);
+
+    public ProdutoDTO adicionar(ProdutoDTO produtoDto) {
+
+        // Remover o id pata conseguir fazer o cadastro
+        produtoDto.setId(null);
+
+        // Criar um objeto de mapeamento
+        ModelMapper mapper = new ModelMapper();
+
+        // Converter o nosso produtoDTO em produto
+        Produto produto = mapper.map(produtoDto, Produto.class);
+
+        // Salvar o produto do banco
+        produto = produtoRepository.save(produto);
+        produtoDto.setId(produto.getId());
+
+        // Retornar o produtoDtO atualizado
+        return produtoDto;
 
     }
 
@@ -52,6 +83,12 @@ public class ProdutoService {
      * @param id do produto a ser deletado.
      */
     public void deletar(Integer id) {
+        Optional<Produto> produto = produtoRepository.findById(id);
+
+        if (produto.isEmpty()) {
+            throw new ResourceNotFoundException(
+                    "Não foi possivel deletar o produto com id: " + id + " - Produto não existe");
+        }
         produtoRepository.deleteById(id);
     }
 
@@ -62,9 +99,19 @@ public class ProdutoService {
      * @param id      do produto.
      * @return Retorna o produto após atualizar a lista.
      */
-    public Produto atualizar(Integer id, Produto produto) {
 
-        produto.setId(id);
-        return produtoRepository.save(produto);
+    public ProdutoDTO atualizar(Integer id, ProdutoDTO produtoDto) {
+
+        // Passar o id para o produtoDTO
+        produtoDto.setId(id);
+        // Criar um objeto de mapeamento
+        ModelMapper mapper = new ModelMapper();
+        // Converte o produtoDTO em um produto.
+        Produto produto = mapper.map(produtoDto, Produto.class);
+        // Atualizar o produto no Banco de Dados
+        produtoRepository.save(produto);
+        // Retornar o produtoDTO atualizado...
+        return produtoDto;
     }
+
 }
